requirements.txt
streamlit>=1.33.0
qrcode[pil]>=7.4
pyshorteners>=1.0.1
Pillow>=10.0.0
